move bootMonitors function to more appropriate location, optimize some imports
