+++ core/src/main/scala/org/broadinstitute/dsde/rawls/Boot.scala
@@ -3,9 +3,8 @@ package org.broadinstitute.dsde.rawls
 import java.io.StringReader
 import java.net.InetAddress
 import java.util.concurrent.TimeUnit
-import javax.naming.directory.AttributeInUseException
 
-import akka.actor.{ActorRef, ActorSystem}
+import akka.actor.ActorSystem
 import akka.io.IO
 import akka.pattern.ask
 import akka.util.Timeout
@@ -13,7 +12,7 @@ import com.codahale.metrics.SharedMetricRegistries
 import com.google.api.client.googleapis.auth.oauth2.GoogleClientSecrets
 import com.google.api.client.json.jackson2.JacksonFactory
 import com.readytalk.metrics.{StatsDReporter, WorkbenchStatsD}
-import com.typesafe.config.{Config, ConfigFactory, ConfigObject, ConfigRenderOptions}
+import com.typesafe.config.{ConfigFactory, ConfigObject}
 import com.typesafe.scalalogging.LazyLogging
 import slick.backend.DatabaseConfig
 import slick.driver.JdbcDriver
@@ -21,25 +20,21 @@ import org.broadinstitute.dsde.rawls.dataaccess._
 import org.broadinstitute.dsde.rawls.dataaccess.jndi.DirectoryConfig
 import org.broadinstitute.dsde.rawls.genomics.GenomicsService
 import org.broadinstitute.dsde.rawls.google.HttpGooglePubSubDAO
-import org.broadinstitute.dsde.rawls.jobexec.{SubmissionSupervisor, WorkflowSubmissionActor}
 import org.broadinstitute.dsde.rawls.model.{ApplicationVersion, UserInfo}
 import org.broadinstitute.dsde.rawls.monitor._
 import org.broadinstitute.dsde.rawls.statistics.StatisticsService
 import org.broadinstitute.dsde.rawls.status.StatusService
 import org.broadinstitute.dsde.rawls.user.UserService
-import org.broadinstitute.dsde.rawls.util._
 import org.broadinstitute.dsde.rawls.util.ScalaConfig._
+import org.broadinstitute.dsde.rawls.util._
 import org.broadinstitute.dsde.rawls.webservice._
 import org.broadinstitute.dsde.rawls.workspace.WorkspaceService
 import spray.can.Http
-import spray.http.StatusCodes
-import spray.json._
 
 import scala.collection.JavaConversions._
-import scala.concurrent.Await
 import scala.concurrent.ExecutionContext.Implicits.global
 import scala.concurrent.duration._
-import scala.util.{Failure, Success, Try}
+import scala.util.{Failure, Success}
 
 object Boot extends App with LazyLogging {
   private def startup(): Unit = {
@@ -173,7 +168,7 @@ object Boot extends App with LazyLogging {
 
     if(conf.getBooleanOption("backRawls").getOrElse(false)) {
       logger.info("This instance has been marked as BACK. Booting monitors...")
-      bootMonitors(
+      BootMonitors.bootMonitors(
         system, conf, slickDataSource, gcsDAO, pubSubDAO, methodRepoDAO, shardedExecutionServiceCluster, maxActiveWorkflowsTotal,
         maxActiveWorkflowsPerUser, bucketDeletionMonitor, userServiceConstructor, projectTemplate, metricsPrefix
       )
@@ -260,58 +255,6 @@ object Boot extends App with LazyLogging {
       .convertDurationsTo(TimeUnit.MILLISECONDS)
       .build(WorkbenchStatsD(host, port))
     reporter.start(period.toMillis, period.toMillis, TimeUnit.MILLISECONDS)
-  }
-
-  def bootMonitors(system: ActorSystem, conf: Config, slickDataSource: SlickDataSource, gcsDAO: HttpGoogleServicesDAO,
-                   pubSubDAO: HttpGooglePubSubDAO, methodRepoDAO: HttpMethodRepoDAO, shardedExecutionServiceCluster: ExecutionServiceCluster,
-                   maxActiveWorkflowsTotal: Int, maxActiveWorkflowsPerUser: Int, bucketDeletionMonitor: ActorRef,
-                   userServiceConstructor: (UserInfo) => UserService, projectTemplate: ProjectTemplate, metricsPrefix: String): Unit = {
-    //TODO: once bucketDeletionMonitor is broken out and db-triggered, it can be handled the same way as the below monitors
-    BootMonitors.restartMonitors(slickDataSource, gcsDAO, bucketDeletionMonitor)
-
-    system.actorOf(CreatingBillingProjectMonitor.props(slickDataSource, gcsDAO, projectTemplate))
-
-    //Boot google group sync monitor
-    val gcsConfig = conf.getConfig("gcs")
-
-    system.actorOf(GoogleGroupSyncMonitorSupervisor.props(
-      util.toScalaDuration(gcsConfig.getDuration("groupMonitor.pollInterval")),
-      util.toScalaDuration(gcsConfig.getDuration("groupMonitor.pollIntervalJitter")),
-      pubSubDAO,
-      gcsConfig.getString("groupMonitor.topicName"),
-      gcsConfig.getString("groupMonitor.subscriptionName"),
-      gcsConfig.getInt("groupMonitor.workerCount"),
-      userServiceConstructor))
-
-    //Boot submission monitor
-    val submissionMonitorConfig = conf.getConfig("submissionmonitor")
-
-    system.actorOf(SubmissionSupervisor.props(
-      shardedExecutionServiceCluster,
-      slickDataSource,
-      gcsDAO.getBucketServiceAccountCredential,
-      util.toScalaDuration(submissionMonitorConfig.getDuration("submissionPollInterval")),
-      submissionMonitorConfig.getBoolean("trackDetailedSubmissionMetrics"),
-      workbenchMetricBaseName = metricsPrefix
-    ).withDispatcher("submission-monitor-dispatcher"), "rawls-submission-supervisor")
-
-    //Boot workflow submission actors
-    for(i <- 0 until conf.getInt("executionservice.parallelSubmitters")) {
-      system.actorOf(WorkflowSubmissionActor.props(
-        slickDataSource,
-        methodRepoDAO,
-        gcsDAO,
-        shardedExecutionServiceCluster,
-        conf.getInt("executionservice.batchSize"),
-        gcsDAO.getBucketServiceAccountCredential,
-        util.toScalaDuration(conf.getDuration("executionservice.processInterval")),
-        util.toScalaDuration(conf.getDuration("executionservice.pollInterval")),
-        maxActiveWorkflowsTotal,
-        maxActiveWorkflowsPerUser,
-        Try(conf.getObject("executionservice.defaultRuntimeOptions").render(ConfigRenderOptions.concise()).parseJson).toOption,
-        workbenchMetricBaseName = metricsPrefix
-      ))
-    }
   }
 
   startup()
